% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinstall.R
\name{chinstall}
\alias{chinstall}
\title{chinstall is the chill way to install and load packages that you may or may
not already have installed in R. It is a terse wrapper around the common and
verbose pattern of:
1. Checking if a package can be loaded
2. Installing if you do not have the required package}
\usage{
chinstall(package_as_string)
}
\arguments{
\item{package_as_string}{dataframe to visualse.}
}
\description{
Created by LD and TPC
}
