% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countThings.R
\name{countThings}
\alias{countThings}
\title{Count Distinct Values in a Dataframe}
\usage{
countThings(data, id_cols, group_cols)
}
\arguments{
\item{data}{A dataframe to process.}

\item{id_cols}{Vector of column names that form the minimal unique identity of interest.}

\item{group_cols}{Vector of column names used for grouping and counting.}
}
\value{
A dataframe with counts of distinct values based on specified groupings.
}
\description{
This function takes a dataframe and performs a count of distinct values based on a minimal description of interest and a grouping for counting.
}
\examples{
data(mtcars)
countThings(mtcars, c("gear", "carb"), c("cyl"))
}
