% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggressR.R
\name{egressR}
\alias{egressR}
\title{takes a dataframe and two variables (a response and predictor),
creates a linear model, and uses this model to predict and normalise the
response variable. The function generates predictions,
and normalises the response variable.
It outputs model visualisation and parameters (R^2 and coefficients)
along the way.}
\usage{
egressR(
  data_frame_to_build_model,
  response_var,
  predictor_variable,
  data_frame_to_apply_model = NULL
)
}
\arguments{
\item{data_frame_to_build_model}{dataframe to visualse.}

\item{response_var}{column with rownumbers.}

\item{predictor_variable}{colNames}

\item{data_frame_to_apply_model}{Number of rows if not 8.#'}
}
\value{
A dataframe with prediction and normalised variables
}
\description{
Function can be applied to the original data frame, or a second one
supplied by the user.
}
