% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastloadr.R
\name{fastloadr}
\alias{fastloadr}
\title{Add together two numbers}
\usage{
fastloadr(csv_path)
}
\arguments{
\item{csv_path}{path to the csv you wish to read as RDS if you can!}
}
\value{
A plot of summarised data
}
\description{
Add together two numbers
}
