% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVertigo.R
\name{ggVertigo}
\alias{ggVertigo}
\title{ggVertigo}
\usage{
ggVertigo(angle = 90)
}
\arguments{
\item{angle}{the angle of rotation.}
}
\value{
returns theme function with arguments to rotate axis labels
}
\description{
Rotates ggplot x axis
}
