% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_emGeezen.R
\name{new_emGeezen}
\alias{new_emGeezen}
\title{Generate a Plate Visualization}
\usage{
new_emGeezen(
  plate_data = NULL,
  variable_to_squiz = NULL,
  low_col = "darkblue",
  high_col = "hotpink",
  rowNames = "Row",
  colNames = "Column",
  summary = "mean",
  numRows = 8,
  numCols = 12,
  discrete = FALSE
)
}
\arguments{
\item{plate_data}{A data frame representing the plate data.}

\item{variable_to_squiz}{The variable of interest to visualize.}

\item{low_col}{The color for the low end of the gradient.}

\item{high_col}{The color for the high end of the gradient.}

\item{rowNames}{The name to be used for rows in the plate visualization.}

\item{colNames}{The name to be used for columns in the plate visualization.}

\item{summary}{The summary statistic to use (default is mean).}

\item{numRows}{The number of rows in the plate.}

\item{numCols}{The number of columns in the plate.}

\item{discrete}{Whether the fill scale is discrete or not.}
}
\value{
A ggplot object representing the plate visualization.
}
\description{
This function takes in plate data and generates a visualization based on the specified parameters.
It allows customization of colors, summary statistics, row and column names, and more.
}
\examples{
new_emGeezen(plate_data = yourData,
             variable_to_squiz = "yourVariable",
             low_col = "darkblue",
             high_col = "hotpink")
}
