% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeTtest.R
\name{perform_paired_t_test}
\alias{perform_paired_t_test}
\title{Perform Paired T-Test and Display Pairing}
\usage{
perform_paired_t_test(data, variable, group, paringID)
}
\arguments{
\item{data}{A dataframe containing the data to be analyzed.}

\item{variable}{The name of the column in \code{data} containing the variable
to test.}

\item{group}{The name of the column in \code{data} that defines the groups
for pairing.}

\item{paringID}{A vector of column names in \code{data} used as identifiers
for pairing.}
}
\value{
A list containing the results of the t-test and a dataframe
  showing the paired data.
}
\description{
This function performs a paired t-test on a specified variable within a
dataframe, grouped by a specified grouping variable. It also creates a
dataframe showing the pairs as used in the test, allowing users to check
the pairing.
}
\examples{
# Assuming 'df' is a dataframe with columns 'relative_change', 'treatment',
# and 'cDNA_template_ng'
result <- perform_paired_t_test(df, "relative_change", "treatment",
                               c("cDNA_template_ng"))
print(result$TestResult)
print(result$PairedData)
}
