% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSure.R
\name{writeSure}
\alias{writeSure}
\title{Write Data to CSV with Assured Directory Creation}
\usage{
writeSure(data, file_path, ...)
}
\arguments{
\item{data}{A data frame, matrix, or other object convertible to CSV format.}

\item{file_path}{The file path where the CSV file will be saved, including both the directory path and
the file name. If the directory does not exist, it will be created.}

\item{...}{Additional arguments to be passed to `write.csv`.}
}
\value{
Invisible NULL. This function is used for its side effect of writing a file.
}
\description{
`writeSure` writes a data frame or similar object to a CSV file, ensuring that the directory for the
file exists. If the directory does not exist, it is created, including any necessary parent directories.
}
\examples{
my_data <- data.frame(column1 = 1:5, column2 = letters[1:5])
writeSure(my_data, "path/to/your/directory/filename.csv")
}
